Changed DefaultUUIDs to be data classes
This change fixes an issue with js implementation where it wasn't delegating to its backend value, but rather another uuid instance that was created.

In terms of `equals` & `hashCode` it looks like the existing functions are doing what a data class does anyway so I decided to go with that.

As for `toString`, I've left these overrides in since data class' `toString` implementation would be of the "DefaultUuid(backend=...)"